feat: add cache id generation for parser
