+++ packages/message-compiler/src/parser.ts
@@ -1,93 +1,25 @@
-import { createLocation, SourceLocation } from './location'
+import { createLocation } from './location'
 import { createCompileError, CompileErrorCodes } from './errors'
 import { createTokenizer, TokenTypes } from './tokenizer'
+import { NodeTypes } from './nodes'
 import { assign } from '@intlify/shared'
 
 import type { Position } from './location'
 import type { ParserOptions } from './options'
 import type { Tokenizer, Token } from './tokenizer'
-
-export const enum NodeTypes {
-  Resource, // 0
-  Plural,
-  Message,
-  Text,
-  Named,
-  List, // 5
-  Linked,
-  LinkedKey,
-  LinkedModifier,
-  Literal
-}
-
-// not containing whitespace or control characters
-export type Identifier = string
-
-export interface Node {
-  type: NodeTypes
-  start?: number
-  end?: number
-  loc?: SourceLocation
-}
-
-export interface ResourceNode extends Node {
-  type: NodeTypes.Resource
-  body: MessageNode | PluralNode
-  helpers?: string[]
-}
-
-export interface PluralNode extends Node {
-  type: NodeTypes.Plural
-  cases: MessageNode[]
-}
-
-export interface MessageNode extends Node {
-  type: NodeTypes.Message
-  items: MessageElementNode[]
-}
-
-type MessageElementNode =
-  | TextNode
-  | NamedNode
-  | ListNode
-  | LiteralNode
-  | LinkedNode
-
-export interface TextNode extends Node {
-  type: NodeTypes.Text
-  value: string
-}
-
-export interface NamedNode extends Node {
-  type: NodeTypes.Named
-  key: Identifier
-}
-
-export interface ListNode extends Node {
-  type: NodeTypes.List
-  index: number
-}
-
-export interface LiteralNode extends Node {
-  type: NodeTypes.Literal
-  value: string
-}
-
-export interface LinkedNode extends Node {
-  type: NodeTypes.Linked
-  modifier?: LinkedModifierNode
-  key: LinkedKeyNode | NamedNode | ListNode | LiteralNode
-}
-
-export interface LinkedKeyNode extends Node {
-  type: NodeTypes.LinkedKey
-  value: string
-}
-
-export interface LinkedModifierNode extends Node {
-  type: NodeTypes.LinkedModifier
-  value: Identifier
-}
+import type {
+  Node,
+  ResourceNode,
+  MessageNode,
+  PluralNode,
+  TextNode,
+  LinkedKeyNode,
+  LinkedNode,
+  ListNode,
+  LiteralNode,
+  LinkedModifierNode,
+  NamedNode
+} from './nodes'
 
 export interface Parser {
   parse(source: string): ResourceNode
@@ -541,6 +473,10 @@ export function createParser(options: ParserOptions = {}): Parser {
       node.loc.source = source
     }
     node.body = parseResource(tokenizer)
+
+    if (options.generateCacheId) {
+      node.cacheId = options.generateCacheId(source)
+    }
 
     // assert whether achieved to EOF
     if (context.currentType !== TokenTypes.EOF) {
