+++ packages/message-compiler/test/parser/literal.test.ts
@@ -1,6 +1,7 @@
-import { createParser, NodeTypes } from '../../src/parser'
+import { createParser } from '../../src/parser'
+import { NodeTypes } from '../../src/nodes'
 
-import type { MessageNode } from '../../src/parser'
+import type { MessageNode } from '../../src/nodes'
 
 let spy: any // eslint-disable-line @typescript-eslint/no-explicit-any
 beforeEach(() => {
